requirements.txt
celery==5.5.3
fastapi[standard]==0.121.0
redis==7.0.1
pydantic_settings==2.11.0
python-dotenv==1.2.1
itsdangerous==2.2.0
